requirements.txt
fastapi
uvicorn
requests
google-generativeai
tiktoken
pytest
httpx 